Add store / button logic
Card buttons now actually work (except for watch later)! Editing, deleting, or adding a movie automatically saves it to a cache file in the user data directory that is initialized on app start up.

Still need to add logic for the "watch later" button. Originally wanted to have it set to plan the dates of movies, but I may experiment with adding a "movie order" of sorts for prioritization rather than scheduling.

Also moved logic into reusable components.